Add more examples to `help use`
The `members` argument of `use` is specified as type `any`,
but it's really a string or list of strings. So add some examples that
mention what you can specify for `members`.

Also mention `help modules` and `help std`, since you probably want to
use the standard library or another defined module.