Update to most recent reedline state
This includes the changes for history exclusion only on the reedline
side, needs explicit enabling on the consumer side.

Also changes the behavior of keybindings with modifiers that previously
where unsupported. Now it is possible to bind keys without a modifier or
that use Ctrl+Alt in directly inserting modes.