use "search_result" style to colorize matching strings (fixes #9275)
This change introduces new `search_result` style supported in the color config. The change also
removes obsolete check for `config.ls_colors` for computing the style. `config.ls_colors` has been
removed last year, so this removes the reference to the obsolete flag, along with a cleanup that
removes all the code that used to rely on ls_colors for highlighting search results.