http post --content-type should set Content-Type header
Looking into https://github.com/nushell/nushell/issues/9408

A few things I found:
The example in the issue does set the `content-type` to `application/json` but sends the body as a string.

If we send the same request but using actual json as the body, the Header is set correctly.