Fix SIGTTIN handling (#9681)
# Description
Due to a typo? in the
[reference](https://www.gnu.org/software/libc/manual/html_node/Initializing-the-Shell.html)
used to implement SIGTTIN handling, nushell will crash when being sent
to the background from within another shell.

For example, currently in bash:
```bash
$ nu -n &
[1] 176058
$ ERROR: failed to SIGTTIN ourselves

[1]+  Exit 1                  nu -n
$
```

Now fixed:
```bash
$ nu -n &
[1] 178788
$ jobs
[1]+  Stopped                 nu -n
$
```

For further reference, this is how
[fish](https://github.com/fish-shell/fish-shell/blob/493cbeb84c9146341790ef563434d47f3b4e0fa5/src/reader.cpp#L2571)
does it.

# User-Facing Changes
Bug fix only -- users should now be able to send nushell to the
background.