Remove `Signature.vectorizes_over_list` entirely.
With the current typechecking logic this property has no effect.
It was only used in the example testing, and provided some indication of
this vectorizing property.
With #9742 all commands that previously declared it have explicit list
signatures. If we want to get it back in the future we can reconstruct
it from the signature.

Simplifies the example testing a bit.

Causes a breaking change for plugins that previously declared it.
This will be a good reminder to update their signature as well to
reflect the more stringent type checking.