Put `explore` command behind `extra` flag
This is a breaking change, to use or run the `explore` command, you
will need to `run`/`build`/`install` with `--features extra`.

This was raised through discussions how far along the `explore` command
is in being 1.0 stable. So far there is still a bit of concerted effort
necessary to polish `explore` so a broad userbase can benefit from it.

This change can easily be reverted if we feel ready to commit to the
`explore` behavior.

This should not interfere with currently ongoing `explore` polishing
happening in #9966, #9964 etc.