Support Termux
On Android, Nushell currently fails to compile because the `pwd` library
attempts to use `getpwent` to fetch a user's home directory from the OS,
and this function does not exist on Android. The Termux
package currently works around this problem with a
[special patch](https://github.com/termux/termux-packages/blob/7f942e3225bde77fd162d2d378e19fe93b895f3c/packages/nushell/user_home_dir.patch).

This commit solves this issue by adding platform conditional code to
fetch the home dir. Additionally, the logic is refactored and combined
for Windows and Mac as well, since all three platforms can benefit from
`dirs_next`, and the fallback path differs only by the hardcoded base
path.

Additionally, when the platform is specifically Termux, the home
directory can be hardcoded altogether, as Termux is single user and
always has the same home directory.